requirements.txt
arxiv==0.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Deprecated==1.2.7
feedparser==5.2.1
idna==2.8
pycparser==2.19
PyGithub==1.45
PyJWT==1.7.1
pytest-runner==5.2
python-dateutil==2.8.1
requests==2.22.0
six==1.14.0
urllib3==1.25.8
wrapt==1.11.2
